PN-14790 WO Task Sequence changes

+++ PAS_Database/dbo/User Defined Types/WorkOrderTasks.sql
@@ -1,8 +1,10 @@
 ﻿CREATE TYPE [dbo].[WorkOrderTasks] AS TABLE (
-    [WorkOrderTaskId] BIGINT        NULL,
-    [WorkOrderId]     BIGINT        NULL,
-    [TaskId]          BIGINT        NULL,
-    [SequenceNumber]  INT           NULL,
-    [MasterCompanyId] INT           NULL,
-    [UpdatedBy]       VARCHAR (256) NULL);
+    [WorkOrderTaskId] BIGINT          NULL,
+    [WorkOrderId]     BIGINT          NULL,
+    [TaskId]          BIGINT          NULL,
+    [SequenceNumber]  DECIMAL (10, 3) NULL,
+    [MasterCompanyId] INT             NULL,
+    [UpdatedBy]       VARCHAR (256)   NULL);
+
+
